requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
python-telegram-bot==20.5
openai==1.0.0

